re-organise documents
