+++ studio/schemas/documents/eu/eu-europe-tab.js
