+++ studio/schemas/documents/instruments/instrument.js
@@ -1,7 +1,7 @@
 import { CgInternal } from 'react-icons/cg';
 import { VscLaw } from 'react-icons/vsc';
 
-import { lineBreak } from '../../components/headerInput';
+import { lineBreak } from '../../../components/headerInput';
 
 export default {
   title: 'instrument',
