+++ studio/schemas/documents/instruments/product-chain.js
