+++ studio/schemas/documents/instruments/simple-thema.js
