+++ studio/schemas/documents/instruments/thema.js
